pull and build cmd/devp2p on every run
The crawler depends on hard fork definitions in go-ethereum, and we'd
need to rebuild the container every time those change. It's easy to
forget that.

This commit changes the script to pull and compile go-ethereum
cmd/devp2p on every crawler run instead of baking the binary into the
container.